Make combined dictionary contain a variable with all definitions
The combined dictionary is a raw JSON array of definitions. Thus,
it cannot be used directly by the Kanken game.

This commit changes the merge script so that it makes the combined
dictionary contain all definitions in a variable `all_questions`.
This way it can be directly included by the game and no further
processing is needed.